Moved js build code to NativeBuild.scala.
